Implement Requested Changes
